member table updated with address and issues in Get enrollment by Mem…

+++ BackEnd/MaxFitGym/MaxFitGym/IRepository/IProgramRepository.cs
@@ -9,7 +9,7 @@ namespace MaxFitGym.IRepository
 
         ICollection<Programs> GetAllPrograms();
 
-        Programs GetProgramById(int ProgramId);
+        Programs GetProgramById(Int64 ProgramId);
 
         void UpdateProgram(int ProgramID, int TotalFee ,string NewName);
